+++ src/constants.ts
@@ -5,3 +5,10 @@ export const ADC_FILE_PATH =
   ".config/gcloud/application_default_credentials.json";
 
 export const CACHE_VERSION = `${APP_NAME}_cache_v1`;
+
+export const WEBVIEW_COMMAND = {
+  switch_config: "switch_config",
+  open_adc_file: "open_adc_file",
+  open_add_config_panel: "open_add_config_panel",
+  create_config: "create_config",
+};