+++ src/helpers.ts
@@ -1,20 +1,47 @@
+import vscode from "vscode";
 import { exec } from "child_process";
 import fs from "fs";
 import os from "os";
 import path from "path";
-import { APPLICATION_DEFAULT_CREDENTIAL, GCP_CONFIGURATION } from "./types";
-import { ADC_FILE_PATH } from "./constants";
+import {
+  APPLICATION_DEFAULT_CREDENTIAL,
+  GCP_CONFIGURATION,
+  NEW_CONFIG_FORM,
+} from "./types";
+import { ADC_FILE_PATH, APP_NAME } from "./constants";
 
-export const activateConfig = async ({ gcpConfig }: any) => {
-  return new Promise((resolve, reject) => {
+export const createConfig = async (newConfig: NEW_CONFIG_FORM) => {
+  return new Promise<NEW_CONFIG_FORM>((resolve, reject) => {
     exec(
-      `gcloud config configurations activate ${gcpConfig.name}`,
+      `gcloud config configurations create ${newConfig.configName} ${
+        !newConfig.activateConfig ? "--no-activate" : ""
+      }`,
       (error, _, stderr) => {
         if (error) {
           reject(error);
           return;
         }
-        if (stderr && !stderr.includes(gcpConfig.name)) {
+        if (stderr && !stderr.includes(newConfig.configName)) {
+          reject(false);
+          return;
+        }
+
+        resolve(newConfig);
+      }
+    );
+  });
+};
+
+export const activateConfig = async (gcpConfigName: string) => {
+  return new Promise((resolve, reject) => {
+    exec(
+      `gcloud config configurations activate ${gcpConfigName}`,
+      (error, _, stderr) => {
+        if (error) {
+          reject(error);
+          return;
+        }
+        if (stderr && !stderr.includes(gcpConfigName)) {
           reject(false);
           return;
         }
@@ -25,10 +52,10 @@ export const activateConfig = async ({ gcpConfig }: any) => {
   });
 };
 
-export const setAccount = async ({ gcpConfig }: any) => {
+export const setAccount = async (gcpConfigAccount: string) => {
   return new Promise((resolve, reject) => {
     exec(
-      `gcloud config set account ${gcpConfig.properties.core.account}`,
+      `gcloud config set account ${gcpConfigAccount}`,
       (error, stdout, stderr) => {
         if (error) {
           reject(error);
@@ -100,7 +127,10 @@ export const updateJsonFile = (
 export const configNameToTitle = (configName = "") =>
   configName.replaceAll("-", " ");
 
-export const createAbsolutePath = (relativePath: string) =>
+export const createOsAbsolutePath = (relativePath: string) =>
+  path.join(os.homedir(), relativePath);
+
+export const createExtensionAbsolutePath = (relativePath: string) =>
   path.join(os.homedir(), relativePath);
 
 export const getGcpConfigurations = async () => {
@@ -130,4 +160,64 @@ export const getGcpConfigurations = async () => {
       }
     );
   });
+};
+
+export const createHtmlHead = (
+  extensionContext: vscode.ExtensionContext,
+  panel: vscode.WebviewPanel
+) => {
+  const codiconsUri = panel.webview.asWebviewUri(
+    vscode.Uri.joinPath(
+      extensionContext.extensionUri,
+      "node_modules",
+      "@vscode/codicons",
+      "dist",
+      "codicon.css"
+    )
+  );
+
+  const stylesUri = panel.webview.asWebviewUri(
+    vscode.Uri.joinPath(extensionContext.extensionUri, "assets", "styles.css")
+  );
+
+  return `
+    <head>
+      <meta charset="UTF-8">
+      <meta http-equiv="Content-Security-Policy" font-src ${panel.webview.cspSource}; style-src ${panel.webview.cspSource};">
+      <meta name="viewport" content="width=device-width, initial-scale=1.0">
+      <title>GCP Switch Configuration</title>
+      <link href="${stylesUri}" rel="stylesheet"/>
+      <link href="${codiconsUri}" rel="stylesheet" />
+    </head>
+  `;
+};
+
+export const createWebViewPanel = (
+  extensionContext: vscode.ExtensionContext,
+  viewColumn: vscode.ViewColumn = vscode.ViewColumn.One
+) => {
+  const panel = vscode.window.createWebviewPanel(
+    APP_NAME,
+    "GCP Switch Config",
+    viewColumn,
+    {
+      enableScripts: true,
+      retainContextWhenHidden: true,
+      enableCommandUris: true,
+      localResourceRoots: [
+        vscode.Uri.joinPath(extensionContext.extensionUri, "assets"),
+        vscode.Uri.joinPath(extensionContext.extensionUri, "node_modules"),
+      ],
+    }
+  );
+
+  const iconPath = vscode.Uri.joinPath(
+    extensionContext.extensionUri,
+    "assets",
+    "gcp-logo.png"
+  );
+
+  panel.iconPath = { dark: iconPath, light: iconPath };
+
+  return panel;
 };