+++ src/views/dashboard.ts
@@ -1,77 +1,33 @@
 import vscode from "vscode";
 import { GCP_CONFIGURATION } from "../types";
-import { configNameToTitle } from "../helpers";
+import { createHtmlHead } from "../helpers";
 import { gcpConfigCard } from "./components/gcp-config-card";
 import { loadingSpinner } from "./components/loading-spinner";
+import { gcpTopbar } from "./components/gco-topbar";
 
 type DashboardProps = {
-  extentionContext: vscode.ExtensionContext;
+  extensionContext: vscode.ExtensionContext;
   gcpConfigurations: GCP_CONFIGURATION[];
-  webview: vscode.Webview;
-  extensionUri: vscode.Uri;
+  panel: vscode.WebviewPanel;
 };
 
 export const dashboardView = ({
-  // extentionContext,
+  extensionContext,
   gcpConfigurations,
-  webview,
-  extensionUri,
+  panel,
 }: DashboardProps) => {
-  const codiconsUri = webview.asWebviewUri(
-    vscode.Uri.joinPath(
-      extensionUri,
-      "node_modules",
-      "@vscode/codicons",
-      "dist",
-      "codicon.css"
-    )
-  );
-
-  const stylesUri = webview.asWebviewUri(
-    vscode.Uri.joinPath(extensionUri, "assets", "styles.css")
-  );
-
-  const gcpActiveConfig = gcpConfigurations.find(
-    (gcpConfig) => gcpConfig.is_active
-  );
-
   return `
           <!DOCTYPE html>
           <html lang="en">
-              <head>
-                  <meta charset="UTF-8">
-                  <meta http-equiv="Content-Security-Policy" font-src ${
-                    webview.cspSource
-                  }; style-src ${webview.cspSource};">
-                  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-                  <title>GCP Switch Configuration</title>
-                  <link href="${stylesUri}" rel="stylesheet"/>
-                  <link href="${codiconsUri}" rel="stylesheet" />
-              </head>
+            ${createHtmlHead(extensionContext, panel)}
               <body>
-                  <div class="header">
-                      <div id="header-title" style="display: flex; justify-content: flex-start; align-items: center; gap: 3em; flex: 1;"> 
-                        <div class="gcp-current-configuration">
-                          <i class="badge"></i>
-                          <h2>${configNameToTitle(gcpActiveConfig?.name)}</h2>
-                        </div>
-                      </div>
-                      <div style="flex: 1;">
-                        <input class="search-input" autocomplete="off" type="search" spellcheck="false" tabindex="0" aria-label="Search for configurations" aria-autocomplete="list" placeholder="Search for configurations">
-                      </div>
-                      <div style="display: flex; justify-content: flex-end; align-items: center; gap: 2em; flex: 1;">
-                        <div class="gcp-notification">
-                          <i class='codicon codicon-bell' style="font-size: 24px;"></i>
-                          <div class="gcp-notification-spinner">
-                          </div>
-                        </div>
-                        <div class="gcp-current-adc">
-                          <i class='codicon codicon-key'></i>
-                          <a href="#" onclick="handleADCJsonClick()">
-                            ADC.json
-                          </a>
-                        </div>
-                      </div>
+                  ${gcpTopbar(gcpConfigurations)}
+                  <div class="gcp-toolbar">
+                    <div>
+                      <button class="button-base button-contained-secondary" onclick="handleOpenAddConfigPanelClick()">
+                        <i class='codicon codicon-add'></i> Configuration
+                      </button>
+                    </div>
                   </div>
                   
                   <div style="padding: 0 24px;">
@@ -96,6 +52,12 @@ export const dashboardView = ({
                       function handleADCJsonClick() {
                         vscode.postMessage({ 
                           command: "open_adc_file" 
+                        })
+                      }
+
+                      function handleOpenAddConfigPanelClick() {
+                        vscode.postMessage({ 
+                          command: "open_add_config_panel" 
                         })
                       }